fix: address go-review findings in daemon framework
CRITICAL:
- Fix data race: staleCache write under RLock -> collect updates under
  RLock, apply under Lock
- Fix ConnectionManager.Client() return type: *Client -> ClientAPI

HIGH:
- Split 17-method SessionProvider into 7 small interfaces (SessionLister,
  SessionMutator, PreviewProvider, SessionActioner, WorktreeProvider,
  RoleSessionProvider, ConnectionAware)
- Fix Unregister backing array leak in MessageRouter
- Remove empty Client struct placeholder
- Add explicit comments justifying lock-free c.local access

MEDIUM:
- UptimeSeconds int -> int64
- Constructor accepts MessageSender interface instead of *MessageRouter

+++ internal/daemon/api.go
@@ -215,7 +215,7 @@ type MsgSessionInfo struct {
 type HealthResponse struct {
 	APIVersion    int    `json:"api_version"`
 	BinaryVersion string `json:"binary_version"`
-	UptimeSeconds int    `json:"uptime_s"`
+	UptimeSeconds int64  `json:"uptime_s"`
 	SessionCount  int    `json:"sessions"`
 }
